+++ app/ante/ante.go
@@ -6,7 +6,7 @@ import (
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
 	authante "github.com/cosmos/cosmos-sdk/x/auth/ante"
 
-	ethante "github.com/evmos/ethermint/app/ante"
+	ethante "github.com/evmos/evmos/v18/app/ante"
 )
 
 // NewAnteHandler returns an ante handler responsible for attempting to route an