+++ app/encoding.go
@@ -1,8 +1,8 @@
 package app
 
 import (
-	"github.com/cosmos/cosmos-sdk/simapp/params"
-	evmenc "github.com/evmos/ethermint/encoding"
+	"cosmossdk.io/simapp/params"
+	evmenc "github.com/evmos/evmos/v18/encoding"
 )
 
 // MakeEncodingConfig creates the EncodingConfig for realio network